most of the main work is done still not operational however

+++ read.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 uint16_t mergeinstruction(uint8_t b1, uint8_t b0);
-int readfile(const char *filename, uint16_t *instructions);
+int readfile(const char *filename, uint8_t *instructions);
 
 #endif
